+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+
+* fix error message in database
+* add config to cut error message for the column `error-msg-length`. Default value: `255`
+
 # 1.1.0
 
 * update tkit-quarkus-log maven dependency